refactor: move template+files for `index` page
to a folder of their own
